Satisfy sol8 casting rule

+++ contracts/CBOR.sol
@@ -53,9 +53,9 @@ library CBOR {
         } else if(value > 0xFFFFFFFFFFFFFFFF) {
             encodeBigNum(buf, uint(value));
         } else if(value >= 0) {
-            encodeFixedNumeric(buf, MAJOR_TYPE_INT, uint64(value));
+            encodeFixedNumeric(buf, MAJOR_TYPE_INT, uint64(uint256(value)));
         } else {
-            encodeFixedNumeric(buf, MAJOR_TYPE_NEGATIVE_INT, uint64(-1 - value));
+            encodeFixedNumeric(buf, MAJOR_TYPE_NEGATIVE_INT, uint64(uint256(-1 - value)));
         }
     }
